requirements.txt
Django~=4.2
djangorestframework~=3.14.0
django-colorfield==0.8.0
django-cors-headers~=3.14.0
Faker~=18.4.0